+++ lib/readline.lua
@@ -33,17 +33,18 @@ local actions =  {
             local i = w.ibar.input
             local text = i.text
             local pos = i.position
-            if text and #text > 1 and pos > 1 then
-                local left, right = string.sub(text, 2, pos), string.sub(text, pos+1)
+            if text and utf8.len(text) > 1 and pos > 1 then
+                local left = string.sub(text, 2, utf8.offset(text, pos))
+                local right = string.sub(text, utf8.offset(text, pos + 1))
                 if not string.find(left, "%s") then
                     left = ""
-                elseif string.find(left, "%w+%s*$") then
-                    left = string.sub(left, 0, string.find(left, "%w+%s*$") - 1)
+                elseif string.find(left, "%S+%s*$") then
+                    left = string.sub(left, 0, string.find(left, "%S+%s*$") - 1)
                 elseif string.find(left, "%W+%s*$") then
                     left = string.sub(left, 0, string.find(left, "%W+%s*$") - 1)
                 end
                 i.text =  string.sub(text, 1, 1) .. left .. right
-                i.position = #left + 1
+                i.position = utf8.len(left) + 1
             end
         end,
         desc = "Delete previous word.",